refactor: streamline post management and enhance UI with published posts list
