+++ apps/app/src/trpc/routers/post.ts
@@ -1,3 +1,4 @@
+import { TRPCError } from '@trpc/server';
 import { z } from 'zod';
 import { createDraftPost, createNewPost } from '@/lib/mutations';
 import {
@@ -66,17 +67,49 @@ export const postRouter = createTRPCRouter({
 
   delete: protectedProcedure
     .input(z.object({ id: z.string() }))
-    .mutation(async ({ input, ctx: { db } }) => {
+    .mutation(async ({ input, ctx: { db, session } }) => {
       const { id } = input;
+
+      const post = await db.post.findUnique({
+        where: { id },
+        select: { authorId: true, workspaceId: true },
+      });
+
+      if (!post) {
+        throw new TRPCError({
+          code: 'BAD_REQUEST',
+          message: 'Post not found.',
+        });
+      }
+
+      if (post.authorId !== session.user.id) {
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
+          message: 'Unauthorized: Only the author can delete this post.',
+        });
+      }
 
       await db.post.delete({
         where: { id },
       });
     }),
+
   resolve: protectedProcedure
     .input(z.object({ id: z.string() }))
     .mutation(async ({ input, ctx: { db, session } }) => {
       const { id } = input;
+
+      const post = await db.post.findUnique({
+        where: { id },
+        select: { resolvedAt: true, workspaceId: true },
+      });
+
+      if (!post) {
+        throw new TRPCError({
+          code: 'BAD_REQUEST',
+          message: 'Post not found.',
+        });
+      }
 
       await db.post.update({
         where: { id },
@@ -86,8 +119,27 @@ export const postRouter = createTRPCRouter({
 
   archive: protectedProcedure
     .input(z.object({ id: z.string() }))
-    .mutation(async ({ input, ctx: { db } }) => {
+    .mutation(async ({ input, ctx: { db, session } }) => {
       const { id } = input;
+
+      const post = await db.post.findUnique({
+        where: { id },
+        select: { authorId: true, archived: true, workspaceId: true },
+      });
+
+      if (!post) {
+        throw new TRPCError({
+          code: 'BAD_REQUEST',
+          message: 'Post not found.',
+        });
+      }
+
+      if (post.authorId !== session.user.id) {
+        throw new TRPCError({
+          code: 'UNAUTHORIZED',
+          message: 'Unauthorized: Only the author can archive this post.',
+        });
+      }
 
       await db.post.update({
         where: { id },
